+++ pkgs/native_assets_cli/test/model/checksum_test.dart
@@ -5,7 +5,6 @@
 import 'dart:io';
 
 import 'package:native_assets_cli/native_assets_cli_internal.dart';
-import 'package:native_assets_cli/src/api/asset.dart';
 import 'package:test/test.dart';
 
 import '../helpers.dart';
@@ -21,10 +20,10 @@ void main() {
       final name1 = HookConfigImpl.checksum(
         packageName: packageName,
         packageRoot: nativeAddUri,
-        targetArchitecture: ArchitectureImpl.x64,
-        targetOS: OSImpl.linux,
-        buildMode: BuildModeImpl.release,
-        linkModePreference: LinkModePreferenceImpl.dynamic,
+        targetArchitecture: Architecture.x64,
+        targetOS: OS.linux,
+        buildMode: BuildMode.release,
+        linkModePreference: LinkModePreference.dynamic,
         supportedAssetTypes: [NativeCodeAsset.type],
         hook: Hook.build,
         version: HookConfigImpl.latestVersion,
@@ -38,10 +37,10 @@ void main() {
       final name2 = HookConfigImpl.checksum(
         packageName: packageName,
         packageRoot: nativeAddUri,
-        targetArchitecture: ArchitectureImpl.x64,
-        targetOS: OSImpl.linux,
-        buildMode: BuildModeImpl.release,
-        linkModePreference: LinkModePreferenceImpl.dynamic,
+        targetArchitecture: Architecture.x64,
+        targetOS: OS.linux,
+        buildMode: BuildMode.release,
+        linkModePreference: LinkModePreference.dynamic,
         dependencyMetadata: {
           'foo': const Metadata({'key': 'value'})
         },
@@ -56,10 +55,10 @@ void main() {
       final name3 = HookConfigImpl.checksum(
         packageName: packageName,
         packageRoot: nativeAddUri,
-        targetArchitecture: ArchitectureImpl.x64,
-        targetOS: OSImpl.linux,
-        buildMode: BuildModeImpl.release,
-        linkModePreference: LinkModePreferenceImpl.dynamic,
+        targetArchitecture: Architecture.x64,
+        targetOS: OS.linux,
+        buildMode: BuildMode.release,
+        linkModePreference: LinkModePreference.dynamic,
         cCompiler: CCompilerConfigImpl(
           compiler: fakeClangUri,
         ),
@@ -74,10 +73,10 @@ void main() {
       final name4 = HookConfigImpl.checksum(
         packageName: packageName,
         packageRoot: nativeAddUri,
-        targetArchitecture: ArchitectureImpl.x64,
-        targetOS: OSImpl.linux,
-        buildMode: BuildModeImpl.release,
-        linkModePreference: LinkModePreferenceImpl.dynamic,
+        targetArchitecture: Architecture.x64,
+        targetOS: OS.linux,
+        buildMode: BuildMode.release,
+        linkModePreference: LinkModePreference.dynamic,
         cCompiler: CCompilerConfigImpl(
           compiler: fakeClangUri,
         ),
@@ -92,10 +91,10 @@ void main() {
       final name5 = HookConfigImpl.checksum(
         packageName: packageName,
         packageRoot: nativeAddUri,
-        targetArchitecture: ArchitectureImpl.x64,
-        targetOS: OSImpl.linux,
-        buildMode: BuildModeImpl.release,
-        linkModePreference: LinkModePreferenceImpl.dynamic,
+        targetArchitecture: Architecture.x64,
+        targetOS: OS.linux,
+        buildMode: BuildMode.release,
+        linkModePreference: LinkModePreference.dynamic,
         supportedAssetTypes: [NativeCodeAsset.type],
         hook: Hook.build,
         version: HookConfigImpl.latestVersion,